requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
pyyaml>=6.0
tqdm>=4.65.0
